Deprecate old join screen
